修改接口问题

+++ src/main/java/com/wx/common/model/request/UserProfileRequest.java
@@ -29,7 +29,7 @@ public class UserProfileRequest {
      */
     private String addr;
 
-    private Integer position;
+    private Long position;
 
     private String source;
